ci/ipsec: Add upgrade/downgrade tests for patch releases
Currently, we test upgrades and downgrades for IPsec against the
previous stable branch, for example:

    - On main branch (v1.15-dev):
            main (PR HEAD) -> v1.14 (branch tip) -> main (PR HEAD)

    - On older stable branches:
            v1.14 (PR HEAD) -> v1.13 (branch tip) -> v1.14 (PR HEAD)

For stable branches, this commit adds support for testing upgrades and
downgrades against the latest patch release as well, for example:

    - On v1.14:
            v1.14 (PR HEAD) -> v1.14.3 (tag) -> v1.14 (PR HEAD)

The workflow currently doesn't work correctly on the main branch, given
that we pass an empty version string when building the directory path to
search for the Helm charts (testing from main branch is covered by the
upgrade/downgrade test to the previous stable branch already). This is
addressed by skipping most of the steps on main branch, in a follow-up
commit.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>